+++ decomp/ST.exe/functions/004DD570/decomp.c
@@ -17,7 +17,7 @@ undefined4 __fastcall FUN_004dd570(TLOBaseTy *param_1)
   }
 LAB_004dd5bb:
   if ((g_manRub3_008117A4 != (STManRub3C *)0x0) && (param_1->field_04D0 == CASE_0)) {
-    iVar1 = thunk_FUN_0062e640(g_manRub3_008117A4,param_1->field_0024,param_1->field_0018);
+    iVar1 = thunk_FUN_0062e640(g_manRub3_008117A4,(int)param_1->field_0024,param_1->field_0018);
     if (iVar1 == 0) {
       if (param_1->field_04D8 == 0) {
         param_1->field_04D8 = 1;
@@ -25,7 +25,7 @@ LAB_004dd5bb:
         return 0;
       }
       if (param_1->field_04DC + 0x2ee <= g_playSystem_00802A38->field_00E4) {
-        iVar1 = thunk_FUN_0062e6b0(g_manRub3_008117A4,param_1->field_0024,param_1->field_0018);
+        iVar1 = thunk_FUN_0062e6b0(g_manRub3_008117A4,(int)param_1->field_0024,param_1->field_0018);
         param_1->field_04D8 = (uint)(iVar1 == 0);
       }
     }
